Update seekbar visibility correctly
If seekbar is configured hidden, it should stay hidden regardless of
the visibility of control bar scrim. Currently we have a list of view
ids in the resource file. These views need to be shown/hidden based
on the scrim visibility, and seek bar is in that list. We can either
ignore seekbar at runtime (this CL does), or delete the seek bar from
the resource file. The latter is hard to mantain because it requires
to configure show_linear_progress_bar and
playback_views_to_hide_when_showing_custom_actions at the same time,
so this CL uses the former solution.

Fixes: 138647453
Test: manual
Change-Id: I0c3d97498f3bf95390ff35131b476be157aecd64