Remove ability to search without the cursor showing up
To reproduce, start searching in Media, hit the done button on the keyboard,
then select the search bar again. The cursor won't reappear, but you'll
still be able to type.

This CL fixes that issue, and also makes it so the search bar is deselected
when clicking the done button.

Test: Manually
Change-Id: Ia7991ae5ed2fd2536e10655bb04e18c1e8e8abfb