Scroll the active item into the screen when the active item changed
This CL updates mActiveItemPos immediately when mActiveQueueItemId
changed, and scrolls the queue to mActiveItemPos.
Previously we update mActiveItemPos in onBindViewHolder(), which
won't be called if the view is not on the screen. So it will fail
if the new active item is not on screen.

To avoid flicker, this CL updates the active item partially instead
of the full item when the progress text changed.

Fixes: 148798081
Test: manual
Change-Id: I215cdaab361ed1d18442c6cbed88a8c0f5f66a9b