Remove custom toolbar back listeners
The toolbar will already call the Activity's onBackPressed()
if there is no back listener that handles it.

This was also needed for rickybobby where we wanted to add
a back button to the top level Media fragment, and pressing
it did nothing, but pressing the android back button did
close MediaCenter.

Test: Manually
Bug: 168153824
Change-Id: Ia213b161197d6de495b4cf181120824636c7e598