Refactor browse UI in preparation for wide screen support
This large change is necessary to support widescreen keyboard and
also to react to parts of the hierarchy becoming obsolete (bluetooth).

Previously, a single recycler view was used to display any list
of media items. Now there is a different view for each list of items.

The bulk of the change consists of only keeping the code to display
a list of media items in BrowseViewController, and move the rest into
the new MediaActivityController which now orchestrates the various
instances of BrowseViewController.


Bug: 162350376 Support widescreen keyboard in Media
Bug: 148811848 BT Switching media sources doesn't refresh browse tree

Fixes: 130760002 Create new BrowseFragments for different tree levels
Fixes: 176906428 Search mode is lost upon UI config changes
Fixes: 131684550 Preserve scroll position when pressing back/up

Fixes: 152785353 Apps are unable to restart after being force killed
Fixes: 123301687 Must recover when the media source dies
Fixes: 145763749 Handle dead media source properly

Test: manual
Change-Id: Id7af43d4fe77c8e10fe9432cbfc50fd5e4689ce9
Merged-In: Id7af43d4fe77c8e10fe9432cbfc50fd5e4689ce9